Refactor 671.
